+++ platforms/src/linux/vulkan_window.cpp
@@ -605,8 +605,8 @@ void WindowVulkanLinux::HandleFocusOut() {
 }
 
 void WindowVulkanLinux::HandleSizeEvent(uint32_t width, uint32_t height) {
-    m_windowCenterX = width / 2;
-    m_windowCenterY = height / 2;
+    m_windowCenterX = static_cast<int>(width) / 2;
+    m_windowCenterY = static_cast<int>(height) / 2;
     m_eventHandler->OnWindowSizeEvent(width, height);
 }
